Dash Collision Crown Drop WIP
Looking into the bug that's causing the player to drop the crown if two unrelated players far away collide during a dash. Still unsure why this is happening -- code seems to be correct upon review, but I need to track down extra controllers to properly replicate and test this bug.