Added Cues to Title Screen and Super Star VO
Cues added for the Announcer to say the title of the game at the start of the main menu.
Added random VO sound cue to a player collecting super star.
Fixed timer VO for the announcer. They were saying the wrong numbers. Works in GM2 but still don't hear it playing in GM1.